Bugfix hyperparameter_tuning

+++ hyperparameter_tuning.py
@@ -108,11 +108,11 @@ def random_search_cv(X_tensor : torch.tensor, y_tensor : torch.tensor, result_fo
             continue  # Skip already computed rows
         weight_decay = row['weight_decay']
         learning_rate = row['learning_rate']
-        hidden_layers = row['hidden_layers']
-        hidden_neurons = row['hidden_neurons']
-        num_epochs = row['num_epochs']
+        hidden_layers = row['hidden_layers'].astype(int)
+        hidden_neurons = row['hidden_neurons'].astype(int)
+        num_epochs = row['num_epochs'].astype(int)
         decrease_factor = row['decrease_factor']
-        fold = row['kFold_id']
+        fold = row['kFold_id'].astype(int)
         
         SEED = 42 + fold + f * k_folds
         train_idx, val_idx = splits[fold]
